Irohad: add subscription engine field
